+++ frontend/src/types/common.ts
@@ -80,34 +80,6 @@ export type Command = {
   run: () => void;
 };
 
-export type EnvironmentQuickActionType =
-  | "quickaction.bb.environment.create"
-  | "quickaction.bb.environment.reorder";
-export type ProjectQuickActionType =
-  | "quickaction.bb.project.create"
-  | "quickaction.bb.project.database.transfer"
-  | "quickaction.bb.project.database.transfer-out";
-export type InstanceQuickActionType = "quickaction.bb.instance.create";
-export type UserQuickActionType = "quickaction.bb.user.manage";
-export type DatabaseQuickActionType =
-  | "quickaction.bb.database.create" // Used by DBA and Owner
-  | "quickaction.bb.database.request" // Used by Developer
-  | "quickaction.bb.database.schema.update"
-  | "quickaction.bb.database.data.update"
-  | "quickaction.bb.database.troubleshoot"
-  | "quickaction.bb.database.schema.sync";
-export type IssueQuickActionType =
-  | "quickaction.bb.issue.grant.request.querier"
-  | "quickaction.bb.issue.grant.request.exporter";
-
-export type QuickActionType =
-  | EnvironmentQuickActionType
-  | ProjectQuickActionType
-  | InstanceQuickActionType
-  | UserQuickActionType
-  | DatabaseQuickActionType
-  | IssueQuickActionType;
-
 export type ResourceType =
   | "PRINCIPAL"
   | "ENVIRONMENT"
